requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.5.0
fastapi==0.109.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.4
mpmath==1.3.0
neo4j==5.16.0
networkx==3.2.1
numpy==1.26.3
openai==1.9.0
packaging==23.2
pinecone-client==3.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
sympy==1.12
tokenizers==0.15.1
torch==2.1.2
tqdm==4.66.1
transformers==4.37.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0
